Multithreading. Task 01. Create class BakAccount
This class is a simple model bank account with synchronized 
method for deposit or withdraw money.
It also override methods toString and equals for using with
thread-safe ADT (abstract data type).
When thread read a file it creates an object  BankAccount with 
name and balance and tried to storing it in ADT. If this object is 
already present in the ADT, then thread get an amount of this object and 
merge it with stored account using method mergeBalance(int amount).